+++ packages/arcgis/bundle/maparcgis/bundle.js
@@ -25,13 +25,13 @@ Oskari.clazz.define("Oskari.arcgis.bundle.arcgis.MapArcGisBundle", function() {
 
 		"scripts" : [{
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/arcgis/maparcgis/plugin/ArcGisLayerPlugin.js"
+			"src" : "../../../../bundles/mapping/maparcgis/plugin/ArcGisLayerPlugin.ol2.js"
 		}, {
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/arcgis/maparcgis/domain/ArcGisLayer.js"
+			"src" : "../../../../bundles/mapping/maparcgis/domain/ArcGisLayer.js"
 		}, {
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/arcgis/maparcgis/domain/ArcGis93Layer.js"
+			"src" : "../../../../bundles/mapping/maparcgis/domain/ArcGis93Layer.js"
 		}
        ]
 	},